add first version

+++ hadoop-itemcount/mapreduce/src/main/java/org/phstudy/ItemCountReducer.java
@@ -1,24 +1,25 @@
 package org.phstudy;
 
 import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
 import java.io.IOException;
 
 
-public class ItemCountReducer extends Reducer<Text, IntWritable, Text, IntWritable> {
+public class ItemCountReducer extends Reducer<Text, LongWritable, Text, LongWritable> {
 
     @Override
     protected void reduce(Text key,
-            Iterable<IntWritable> values,
+            Iterable<LongWritable> values,
             Context context)
             throws IOException, InterruptedException
     {
         int count = 0;
-        for(IntWritable value : values) {
+        for(LongWritable value : values) {
             count += value.get();
         }
-        context.write(key, new IntWritable(count));
+        context.write(key, new LongWritable(count));
     }    
 }
